+++ src/poly.h
@@ -5,13 +5,13 @@
 
 // Polynomial.
 typedef struct {
-  uint8_t deg;    // Degree of polynomial.
+  int8_t deg;    // Degree of polynomial.
   int8_t *coeff;  // Array of coefficients.
-  size_t len;     // Length of the array of coefficients.
+  int8_t len;     // Length of the array of coefficients.
 } poly_t;
 
 // Initialize a polynomial.
-poly_t *poly_from_array(uint8_t deg, int8_t *coeff, size_t len);
+poly_t *poly_from_array(int8_t deg, int8_t *coeff, int8_t len);
 
 // Destroy a given polynomial.
 void poly_destroy(poly_t *a);
@@ -22,7 +22,7 @@ void poly_destroy(poly_t *a);
 int poly_eq(const poly_t *a, const poly_t *b);
 
 /* Return a zero polynomial of the given length.*/
-poly_t *poly_create_zero(size_t len);
+poly_t *poly_create_zero(int8_t len);
 
 /* Set a = a mod b, where a and b are polynomials over Fp.
    Assume deg a >= deg b. */
